requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0

opencv-python>=4.7.0
pillow>=9.0.0
albumentations>=1.3.0

pytorch-metric-learning>=2.0.0

matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0
tqdm>=4.65.0

pyyaml
requests
